LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/invalid-uncompiled-raw-handler-compiled-late.window-expected.txt

CONSOLE MESSAGE: line 8: SyntaxError: Parser error
CONSOLE MESSAGE: line 14: SyntaxError: Parser error

FAIL Invalid uncompiled raw handlers should only be compiled when about to call them assert_array_equals: lengths differ, expected 3 got 4
